Removed reotrofitgenerator and replaced it to api service module as a di singleton
