+++ app/src/main/java/com/example/taskplanner/network/AuthInterceptor.kt
@@ -1,10 +1,17 @@
 package com.example.taskplanner.network
 
+import com.example.taskplanner.storage.Storage
 import okhttp3.Interceptor
 import okhttp3.Response
 
-class AuthInterceptor :Interceptor {
+class AuthInterceptor(private val storage: Storage):Interceptor {
+
     override fun intercept(chain: Interceptor.Chain): Response {
-        TODO("Not yet implemented")
+        val request = chain.request().newBuilder()
+        val token = storage.getToken()
+        if (token.isNotEmpty()){
+            request.addHeader("Authorization", "Bearer $token")
+        }
+        return chain.proceed(request.build())
     }
 }
